+++ Emulator/MMU.hpp
@@ -30,9 +30,9 @@ public:
 
 	void dump();
 
-	virtual uint8_t read8(uint16_t addr) override;
-	virtual void write8(uint16_t addr, uint8_t v) override;
-	virtual bool isInRange(uint16_t addr) const override;
+	virtual uint8_t read8(size_t addr) override;
+	virtual void write8(size_t addr, uint8_t v) override;
+	virtual bool isInRange(size_t addr) const override;
 	virtual size_t size() const override;
 	virtual const char* name() const override;
